[kaneton/machine]   o Various octane.mips64 files updated and added.
[environment/profile/host/linux/ia32.mips64]
  o mips64.mk and mips64.conf updated.

[kaneton/libc/include/limits.h]
  o typeof directive changed to __typeof__

+++ kaneton/machine/architecture/mips64/include/types.h
@@ -5,10 +5,14 @@
  *
  * license       kaneton
  *
- * file          /home/mycure/kane...include/arch/ibm-pc.ia32-virtual/types.h
+ * file          /home/lec_l/kanet...hine/architecture/mips64/include/types.h
  *
- * created       julien quintard   [wed jun  6 19:07:02 2007]
- * updated       julien quintard   [thu jun  7 14:26:38 2007]
+ * created       laurent lec   [fri mar 28 17:09:51 2008]
+ * updated       laurent lec   [fri mar 28 17:16:27 2008]
+ */
+
+/*
+ * ---------- information -----------------------------------------------------
  */
 
 #ifndef ARCHITECTURE_TYPES_H
@@ -24,7 +28,7 @@ typedef unsigned short		t_uint16;
 typedef signed short		t_sint16;
 typedef unsigned int		t_uint32;
 typedef signed int		t_sint32;
-typedef unsigned long long int	t_uint64;
-typedef signed long long int	t_sint64;
+typedef unsigned long int	t_uint64;
+typedef signed long int		t_sint64;
 
 #endif
